Add border class to search component and update project dependencies
Added a border class to the Search component to enhance its appearance. Also updated the `pnpm-lock.yaml` file to include recent dependency updates. These updates include the addition of '@emotion/react', '@emotion/styled', '@mui/material', and other related libraries, to broaden the scope of UI components used in the app and provide a more efficient and effective styling solution. All changes were made to align the project with the latest tech standards and to improve the application's overall aesthetics.