Add react-use-websocket library and implement WebSocket in the dashboard
The commit introduces react-use-websocket library to allow effective use of WebSockets. This addition is visible in the 'package.json' and 'pnpm-lock.yml' files. The library is used in the 'Dashboard' component, to manage the WebSocket connection and handle sending and receiving of messages in real-time. 'useWebSocket' Hook is adopted for WebSocket operations. State functionality is adjusted with respective useEffects  and useState hooks to manipulate message history and WebSocket's connection status. Real-time updates in the dashboard are now achievable by refreshing the WebSocket URL. The commit enhances the live interaction capability by integrating a new event listener to the dashboard's textarea for responding to WebSocket messages.