Implement file upload modal in UploadButton component
Enhancements were made to create a file upload modal in the UploadButton component for a better user interaction. This change includes additional functions to manage the UI interactively with state control. Moreover, minor amendments include the ability for the 'setFile' action in 'useMessages' hook to accept null for deselection, and an unneeded import was removed from the dashboard page.