Remove Prisma setup and rename ContextProvider
Deleted Prisma migrations, schema, and lock files as part of phasing out the use of Prisma. Also, renamed ContextProvider to WebSocketContextProvider to provide a more specific context name. Modified codebase and dependencies to accommodate these changes. The changes are aimed at codebase organization and making the purpose of the provider more intuitive.