Add Google Maps functionality to project
Added the Google Maps functionality to the project through the integration of @react-google-maps/api package and enhancements in various files. Implemented a new modal to show the map and ability to add a marker on click. Furthermore, a new GeoButton component was created to open this modal. Magnetic sensor technology was also used to get the user's current location to set the map center.