Add geolocation support to UploadButton component
The UploadButton component now includes geolocation data in the uploaded data and displays a button for setting the geolocation. This is achieved by using a useRef to hold the geolocation data and by adding a GeoButton and MapPinIcon. The geolocation data is sent to the server along with the rest of the form's data in the handleFormSubmit function.