Implement delete message functionality and update Spinner component
A delete button has been added to let users remove messages from the server. Also, an action was defined to manage this operation in the global state. Moreover, Spinner component has been updated for better customizable dimensions and stroke width. This delivers more flexibility for the Spinner usage in different parts of the application.