Enhance UI for login form and home page
Several changes were made to enhance the user interface, focusing on the login form and the home page. An icon was added to the email input field on the login form, and messages were updated to be more engaging. The login button was reformatted to improve visual interest with a hover effect added. A gradient text class was also added to the global CSS and was used on the home page for the main header text.